+++ src/pygpt_net/provider/model/patch.py
@@ -55,9 +55,9 @@ class Patch:
                 data = self.window.core.models.items
                 updated = True
 
-            # < 2.0.104  <--- patch for llama-index gpt4-turbo
-            if old < parse_version("2.0.104"):
-                print("Migrating models from < 2.0.104...")
+            # < 2.0.105  <--- patch for llama-index gpt4-turbo
+            if old < parse_version("2.0.105"):
+                print("Migrating models from < 2.0.105...")
                 self.window.core.updater.patch_file('models.json', True)  # force replace file
                 self.window.core.models.load()
                 data = self.window.core.models.items